Several changes included in this commit...
1.  Restructured (again) how js/css files are loaded in the project.
The jQuery js file is now the only file that loads synchronously.  This
allows the stylesheet information (css.php) to be loaded and get
applied to forms more quickly.  Most all javascript (js.php) is now
loaded asynchonously instead of through a standard <script> tag which
locks the browser while it loads.

2.  The padding and border for textboxes/textarea are now used to adjust
the element's total width to ensure its width does not expand outside of
the appropriate container.  Previously, a 2px padding and a 1px border
were assumed (and hardcoded into the project).  You can demo this
changes in any one of the example files included in this project.
Padding has been set to 5px and border to 1px.

3.  Minor updates to documentation.


git-svn-id: http://php-form-builder-class.googlecode.com/svn/trunk@316 18c9a15c-d092-11de-bb02-e302c5db49d5